requirements.txt
dj-database-url
Django
django-admin-charts
django-admin-rangefilter
django-admin-sso
django-bower
django-classy-tags
django-cms
django-codemirror2
django-cors-headers
django-crispy-forms
django-csp
django-cte
django-evolution
django-extensions
django-filer
django-filter
django-firebase-auth
django-formtools
django-jazzmin
django-js-asset
django-jsonfield
django-material
django-material-admin
django-modelcluster
django-mptt
django-nested-inline
django-nvd3
django-phone-verify
django-phonenumber-field
django-polymorphic
django-queryset-csv
django-river
django-sekizai
django-taggit
django-taggit-anywhere
django-taggit-labels
django-treebeard
djangocms-admin-style
djangocms-attributes-field
djangocms-bootstrap4
djangocms-file
djangocms-googlemap
djangocms-icon
djangocms-installer
djangocms-link
djangocms-picture
djangocms-style
djangocms-text-ckeditor
djangocms-video
djangorestframework
djangorestframework-api-key
djcacheutils
firebase-admin
Flask
flask-ngrok
future
gevent
gevent-websocket
Glymur
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
h5py
html5lib
httplib2
idna
idna-ssl
image-processing
imagecodecs
imagecodecs-lite
imageio
immutables
importlib-metadata
importlib-resources
ipykernel
ipython
ipython-genutils
itsdangerous
jdcal
Jinja2
joblib
jpylyzer
json5
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab-pygments
kiwisolver
l18n
lxml
Markdown
MarkupSafe
matplotlib
mistune
mock
msgpack
multidict
mysql-connector
mysql-connector-python
mysqlclient
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nibabel
notebook
numpy
oauth2client
olefile
opencv-python
openpyxl
openslide-python
packaging
pandas
pandocfilters
pexpect
phonenumbers
pickleshare
Pillow
prometheus-client
prompt-toolkit
proto-plus
protobuf
ptyprocess
pyasn1
pyasn1-modules
pycodestyle
pycparser
pydicom
pyee
Pygments
pyinstaller
pyinstaller-hooks-contrib
PyJWT
pyparsing
PyPDF2
pyrsistent
python-dateutil
python-dotenv
python-memcached
python-nvd3
python-slugify
pytz
PyWavelets
pyzmq
requests
river-admin
rsa
scikit-image
scipy
Send2Trash
simplegeneric
six
sniffio
soupsieve
sqlparse
tablib
terminado
testpath
threadpoolctl
tifffile
toml
toolz
tornado
tqdm
traitlets
psycopg2
